Create baxter-robot
Asks for user input, following the defined rules at the top and depending on their input, appends keywords to the empty array. At the end, the program checks the empty array matches one of the defined syntax rule formats for the array. If it does - success! or it'll fail and tell the user to try again.